fix: modified daily activity input UI design
