+++ app/src/main/AndroidManifest.xml
@@ -21,12 +21,6 @@
             android:name=".ui.eco_balance.Payment.CheckoutActivity"
             android:exported="false" />
         <activity
-            android:name=".LogHabitActivity"
-            android:exported="false" />
-        <activity
-            android:name=".HabitSelectionActivity"
-            android:exported="false" />
-        <activity
             android:name=".LoginActivity"
             android:exported="false" />
         <activity
