+++ app/src/main/java/com/example/planetze/ui/eco_tracker/activities/BuyNewClothesFragment.java
@@ -1,4 +1,4 @@
-package com.example.planetze.ui.eco_tracker;
+package com.example.planetze.ui.eco_tracker.activities;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -12,19 +12,19 @@ import androidx.navigation.NavController;
 import androidx.navigation.fragment.NavHostFragment;
 
 import com.example.planetze.R;
-import com.example.planetze.classes.EcoTracker.Category.Transportation.CyclingOrWalking;
+import com.example.planetze.classes.EcoTracker.Category.Consumption.BuyClothes;
 import com.example.planetze.classes.EcoTracker.Date;
-import com.example.planetze.databinding.FragmentCyclingOrWalkingBinding;
+import com.example.planetze.databinding.FragmentBuyNewClothesBinding;
 
-public class CyclingOrWalkingFragment extends BaseActivityFragment {
+public class BuyNewClothesFragment extends BaseActivityFragment {
 
-    private FragmentCyclingOrWalkingBinding binding;
-    private double distance;
+    private FragmentBuyNewClothesBinding binding;
+    private int num;
 
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        binding = FragmentCyclingOrWalkingBinding.inflate(inflater, container, false);
+        binding = FragmentBuyNewClothesBinding.inflate(inflater, container, false);
         View view = binding.getRoot();
 
         binding.submit.setOnClickListener(this::handleNextButtonClick);
@@ -34,15 +34,15 @@ public class CyclingOrWalkingFragment extends BaseActivityFragment {
 
     private void handleNextButtonClick(View view) {
         try {
-            distance = Double.parseDouble(binding.distance.getText().toString());
+            num = Integer.parseInt(binding.input.getText().toString());
         } catch (Exception e) {
-            Toast.makeText(getActivity(), "Please enter a valid distance", Toast.LENGTH_SHORT).show();
+            Toast.makeText(getActivity(), "Please enter a valid number of clothes", Toast.LENGTH_SHORT).show();
         }
-        if (distance <= 0) {
-            Toast.makeText(getActivity(), "Please enter a valid distance", Toast.LENGTH_SHORT).show();
+        if (num <= 0) {
+            Toast.makeText(getActivity(), "Please enter a valid number of clothes", Toast.LENGTH_SHORT).show();
         } else {
             Date date = Date.today();
-            CyclingOrWalking activity = new CyclingOrWalking(distance);
+            BuyClothes activity = new BuyClothes(num);
             currentUser.addActivity(date, activity);
             databaseManager.add(currentUser);
 
@@ -50,4 +50,5 @@ public class CyclingOrWalkingFragment extends BaseActivityFragment {
             navController.navigate(R.id.eco_tracker);
         }
     }
+
 }
