+++ app/src/main/java/com/example/planetze/ui/eco_tracker/activities/MealFragment.java
@@ -1,4 +1,4 @@
-package com.example.planetze.ui.eco_tracker;
+package com.example.planetze.ui.eco_tracker.activities;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
