+++ app/src/main/java/com/example/planetze/ui/eco_tracker/activities/OtherPurchasesFragment.java
@@ -1,4 +1,4 @@
-package com.example.planetze.ui.eco_tracker;
+package com.example.planetze.ui.eco_tracker.activities;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -12,19 +12,20 @@ import androidx.navigation.NavController;
 import androidx.navigation.fragment.NavHostFragment;
 
 import com.example.planetze.R;
-import com.example.planetze.classes.EcoTracker.Category.Consumption.BuyClothes;
+import com.example.planetze.classes.EcoTracker.Category.Consumption.BuyOthers;
 import com.example.planetze.classes.EcoTracker.Date;
-import com.example.planetze.databinding.FragmentBuyNewClothesBinding;
+import com.example.planetze.databinding.FragmentOtherPurchasesBinding;
 
-public class BuyNewClothesFragment extends BaseActivityFragment {
+public class OtherPurchasesFragment extends BaseActivityFragment {
 
-    private FragmentBuyNewClothesBinding binding;
+    private FragmentOtherPurchasesBinding binding;
+    private String type;
     private int num;
 
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        binding = FragmentBuyNewClothesBinding.inflate(inflater, container, false);
+        binding = FragmentOtherPurchasesBinding.inflate(inflater, container, false);
         View view = binding.getRoot();
 
         binding.submit.setOnClickListener(this::handleNextButtonClick);
@@ -34,15 +35,20 @@ public class BuyNewClothesFragment extends BaseActivityFragment {
 
     private void handleNextButtonClick(View view) {
         try {
-            num = Integer.parseInt(binding.input.getText().toString());
+            num = Integer.parseInt(binding.num.getText().toString());
         } catch (Exception e) {
-            Toast.makeText(getActivity(), "Please enter a valid number of clothes", Toast.LENGTH_SHORT).show();
+            Toast.makeText(getActivity(), "Please enter a valid number of purchases", Toast.LENGTH_SHORT).show();
+        }
+        try {
+            type = binding.type.getText().toString();
+        } catch (Exception e) {
+            Toast.makeText(getActivity(), "Please enter a valid type of purchase", Toast.LENGTH_SHORT).show();
         }
         if (num <= 0) {
-            Toast.makeText(getActivity(), "Please enter a valid number of clothes", Toast.LENGTH_SHORT).show();
+            Toast.makeText(getActivity(), "Please enter a valid number of purchases", Toast.LENGTH_SHORT).show();
         } else {
             Date date = Date.today();
-            BuyClothes activity = new BuyClothes(num);
+            BuyOthers activity = new BuyOthers(type, num);
             currentUser.addActivity(date, activity);
             databaseManager.add(currentUser);
