+++ app/src/main/java/com/example/planetze/ui/eco_tracker/activities/TakePublicTransportationFragment.java
@@ -1,4 +1,4 @@
-package com.example.planetze.ui.eco_tracker;
+package com.example.planetze.ui.eco_tracker.activities;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -13,24 +13,24 @@ import androidx.navigation.NavController;
 import androidx.navigation.fragment.NavHostFragment;
 
 import com.example.planetze.R;
-import com.example.planetze.classes.EcoTracker.Category.Consumption.EnergyBill;
+import com.example.planetze.classes.EcoTracker.Category.Transportation.TakePublicTransportation;
 import com.example.planetze.classes.EcoTracker.Date;
-import com.example.planetze.databinding.FragmentEnergyBillsBinding;
+import com.example.planetze.databinding.FragmentTakePublicTransportationBinding;
 
 import java.util.Arrays;
 import java.util.List;
 
-public class EnergyBillsFragment extends BaseActivityFragment {
+public class TakePublicTransportationFragment extends BaseActivityFragment {
 
-    private FragmentEnergyBillsBinding binding;
+    private FragmentTakePublicTransportationBinding binding;
     private List<Button> buttons;
     private String type;
-    private double amount;
+    private int hour;
 
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        binding = FragmentEnergyBillsBinding.inflate(inflater, container, false);
+        binding = FragmentTakePublicTransportationBinding.inflate(inflater, container, false);
         View view = binding.getRoot();
 
         setOnClickListeners();
@@ -58,15 +58,15 @@ public class EnergyBillsFragment extends BaseActivityFragment {
 
     private void handleNextButtonClick(View view) {
         try {
-            amount = Double.parseDouble(binding.amount.getText().toString());
+            hour = Integer.parseInt(binding.hour.getText().toString());
         } catch (Exception e) {
-            Toast.makeText(getActivity(), "Please enter a valid bill amount", Toast.LENGTH_SHORT).show();
+            Toast.makeText(getActivity(), "Please enter a valid number of hours", Toast.LENGTH_SHORT).show();
         }
-        if (amount <= 0) {
-            Toast.makeText(getActivity(), "Please enter a valid bill amount", Toast.LENGTH_SHORT).show();
+        if (hour <= 0) {
+            Toast.makeText(getActivity(), "Please enter a valid number of hours", Toast.LENGTH_SHORT).show();
         } else {
             Date date = Date.today();
-            EnergyBill activity = new EnergyBill(type, amount);
+            TakePublicTransportation activity = new TakePublicTransportation(type, hour);
             currentUser.addActivity(date, activity);
             databaseManager.add(currentUser);
 
@@ -74,5 +74,4 @@ public class EnergyBillsFragment extends BaseActivityFragment {
             navController.navigate(R.id.eco_tracker);
         }
     }
-
 }
