+++ app/src/main/res/values/strings.xml
@@ -251,5 +251,6 @@
     <string name="profile">Profile</string>
     <string name="user_details">User details</string>
     <string name="add_activity">Add Activity</string>
+    <string name="select_a_project">Select A Project</string>
 
 </resources>
